backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.8.1
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
alembic==1.15.2
numpy==2.2.4
pillow==11.2.1
opencv-python-headless==4.11.0.86
scikit-learn==1.6.1
joblib==1.4.2
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.28.1
